Project backup command added

+++ console/DbBackupCommand.php
@@ -6,9 +6,11 @@ use Illuminate\Support\Facades\Storage;
 
 class DbBackupCommand extends Command
 {
-    protected $signature = 'db:backup {cloud? : The name of the cloud storage (default: dropbox)}';
+    protected $signature = 'db:backup
+        {cloud? : The name of the cloud storage where backup will be uploaded}
+        {--d|--nodelete : Do not delete the backup file after it is uploaded to the cloud storage}';
 
-    protected $description = 'Create database backup and upload it to the cloud storage.';
+    protected $description = 'Create database backup and optionally upload it to the cloud storage.';
 
     const BACKUP_DIRECTORY = 'database';
 
@@ -19,12 +21,10 @@ class DbBackupCommand extends Command
         $dbPass = config("database.connections.{$connection}.password");
         $dbName = config("database.connections.{$connection}.database");
 
-        $this->line('');
-
-        $this->question('Creating database dump file...');
-
         $backupName = Carbon::now()->format('Y-m-d_H-i-s') . '.sql.gz';
 
+        $this->line('');
+        $this->question('Creating database dump file...');
         $this->info(shell_exec("mysqldump -u{$dbUser} -p{$dbPass} {$dbName} | gzip > {$backupName}"));
         $this->info('Database dump file successfully created.');
         $this->line('');
@@ -32,19 +32,17 @@ class DbBackupCommand extends Command
         if ($this->argument('cloud')) {
             $cloudStorage = Storage::disk($this->argument('cloud'));
 
-            if (!$cloudStorage->exists(static::BACKUP_DIRECTORY)) {
-                $cloudStorage->makeDirectory(static::BACKUP_DIRECTORY);
-            }
-
             $this->question('Uploading database dump file to the cloud storage...');
             $cloudStorage->put(static::BACKUP_DIRECTORY . '/' . $backupName, file_get_contents($backupName));
             $this->info('Database dump file successfully uploaded.');
-
             $this->line('');
 
-            $this->question('Deleting the database dump file...');
-            $this->info(shell_exec("rm -f {$backupName}"));
-            $this->info('Database dump file successfully deleted.');
+            if (!$this->option('nodelete')) {
+                $this->question('Deleting the database dump file...');
+                $this->info(shell_exec("rm -f {$backupName}"));
+                $this->info('Database dump file successfully deleted.');
+                $this->line('');
+            }
         }
 
         $this->alert('Database backup was successfully created.');
